working CLI prototype
This prototype if for mp4 videos only GUI.
The 'choose mp4 or mp3' part will be implemented in the future.

in order for it to work you need to manually change the url inside the program file. it works for youtube and pinterest videos. if the youtube video is part of a playlist ,the url will show it and you need to remove the playlist part from the video url in order for it to work. else it will just do nothing. if the internet connection is slow the download will drop. if you download the same video twice or more , the new video is just going to overwrite it.
its going to create  a log file where the program is located and a downloaded_videos_yt folder on desktop where the downloads will be stored.