GUI done. no progress bar. works fine for yt/pinterest
The GUI for the downloader is functional. it does not have progress bars.
if the user leaves the input box empty or enters an invalid url , a QmessageWindow pops up  telling the user about the invalid url.
when a video is successfully downloaded a QMessageWindow pops up telling the user 'yey video downloaded !' , and it has a flower icon.
when you press on the search button it will turn orange , letting the user know that they pressed the button. during the download process , the button gets disabled , that means the user cannot click on it anymore until the download is over or an error occurs, such as invalid url.
if for some reason the internet connection drops  while the video/audio is downloading , the process will be halted.
the moment the wifi comes back on , you will get a nasty traceback OR errors in the final downloaded video.
if the internet connection drops while the video is being downloaded and then it comes back , the audio will be downloaded fine and in the end result you will have a file with perfectly good audio but a freezed video.
ATTENTION!
● when downloading tik toks , it gives a traceback which stops the program. the search button remains disabled and orange and that basically means you have to close the program and open it again. it does this every time with tiktok even if i am loged in. so yea , tik tok is a big nono.
instagram and other platforms apart from youtube/pinterest/tik tok have not been tested.